// services/react-frontend/src/library/common/types/projectTypes.ts
export enum ProjectTypes {
	CREATE_PROJECT = '@@PROJECT/CREATE_PROJECT',
	CREATE_PROJECT_SUCCESS = '@@PROJECT/CREATE_PROJECT_SUCCESS',
	REQUEST_PROJECTS = '@@PROJECT/REQUEST_PROJECTS',
	REQUEST_PROJECTS_SUCCESS = '@@PROJECT/REQUEST_PROJECTS_SUCCESS',
	REQUEST_MEASUREMENTS = '@@PROJECT/REQUEST_MEASUREMENTS',
	REQUEST_MEASUREMENTS_SUCCESS = '@@PROJECT/REQUEST_MEASUREMENTS_SUCCESS',
	REQUEST_PROJECT_BY_ID = '@@PROJECT/REQUEST_PROJECT_BY_ID',
	REQUEST_PROJECT_BY_ID_SUCCESS = '@@PROJECT/REQUEST_PROJECT_BY_ID_SUCCESS',
	EDIT_PROJECT_NAME = '@@PROJECT/EDIT_PROJECT_NAME',
	EDIT_PROJECT_NAME_SUCCESS = '@@PROJECT/EDIT_PROJECT_NAME_SUCCESS',
	DELETE_PROJECT = '@@PROJECT/DELETE_PROJECT',
	DELETE_PROJECT_SUCCESS = '@@PROJECT/DELETE_PROJECT_SUCCESS',
}